test(typespec-ts): move rlc-common unit tests to mirror new homes
Relocate the former test-next/unit/rlc-common tests alongside the code they exercise: helper tests to test-next/unit/utils and metadata/test builder tests to test-next/unit/metadata.
