+++ packages/typespec-ts/test-next/unit/metadata/static/package.json
