+++ packages/typespec-ts/test-next/unit/utils/imports-util.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from "vitest";
 import {
   buildRuntimeImports,
   getImportSpecifier,
-} from "../../../../src/utils/imports-util.js";
+} from "../../../src/utils/imports-util.js";
 
 describe("#buildRuntimeImports", () => {
   it("should return the correct import set", () => {
