+++ packages/typespec-ts/test-next/unit/utils/type-util.test.ts
@@ -17,7 +17,7 @@ import {
   leaveStringQuotes,
   toTypeScriptTypeFromName,
   toTypeScriptTypeFromSchema,
-} from "../../../../src/utils/type-util.js";
+} from "../../../src/utils/type-util.js";
 
 describe("#isStringLiteral", () => {
   it("should return true if the string is quoted", () => {
